requirements.txt
asgiref==3.3.1
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-filter==2.4.0
django==3.1.4
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
faker==4.18.0
markdown==3.3.3
pyjwt==1.7.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
